Move to use application bootstrap
