+++ IsraelHiking.Web/src/application/components/map/main-map.component.ts
@@ -1,7 +1,9 @@
 import { Component, ViewEncapsulation, ElementRef, inject, viewChild, viewChildren } from "@angular/core";
 import { MatDialog, MatDialogConfig } from "@angular/material/dialog";
+import { NgStyle, NgIf } from "@angular/common";
 import { MapComponent, CustomControl } from "@maplibre/ngx-maplibre-gl";
 import { setRTLTextPlugin, StyleSpecification, ScaleControl, Unit, RasterDEMSourceSpecification, PointLike } from "maplibre-gl";
+import { NgProgressComponent } from "ngx-progressbar";
 import { Store } from "@ngxs/store";
 
 import { TracesDialogComponent } from "../dialogs/traces-dialog.component";
@@ -12,13 +14,27 @@ import { RunningContextService } from "../../services/running-context.service";
 import { DefaultStyleService } from "../../services/default-style.service";
 import { SetLocationAction } from "../../reducers/location.reducer";
 import type { ApplicationState, LocationState } from "../../models/models";
+import { SidebarComponent } from "../sidebar/sidebar.component";
+import { BackgroundTextComponent } from "../background-text.component";
+import { LayersViewComponent } from "./layers-view.component";
+import { RoutesComponent } from "./routes.component";
+import { RecordedRouteComponent } from "./recorded-route.component";
+import { TracesComponent } from "./traces.component";
+import { ZoomComponent } from "../zoom.component";
+import { LocationComponent } from "../location.component";
+import { MainMenuComponent } from "../main-menu.component";
+import { SearchComponent } from "../search.component";
+import { DrawingComponent } from "../drawing.component";
+import { RouteStatisticsComponent } from "../route-statistics.component";
+import { CenterMeComponent } from "../center-me.component";
+import { IhmLinkComponent } from "../ihm-link.component";
 
 @Component({
     selector: "main-map",
     templateUrl: "./main-map.component.html",
     styleUrls: ["./main-map.component.scss"],
     encapsulation: ViewEncapsulation.None,
-    standalone: false
+    imports: [NgProgressComponent, NgStyle, SidebarComponent, BackgroundTextComponent, MapComponent, LayersViewComponent, RoutesComponent, RecordedRouteComponent, TracesComponent, ZoomComponent, NgIf, LocationComponent, MainMenuComponent, SearchComponent, DrawingComponent, RouteStatisticsComponent, CenterMeComponent, IhmLinkComponent]
 })
 export class MainMapComponent {
