+++ IsraelHiking.Web/src/application/services/application-initialize.service.ts
@@ -27,6 +27,7 @@ import { ToastService } from "./toast.service";
 import { ApplicationUpdateService } from "./application-update.service";
 import { LocationService } from "./location.service";
 import { HashService } from "./hash.service";
+import { Angulartics2GoogleGlobalSiteTag } from "angulartics2";
 import type { ApplicationState } from "../models/models";
 
 @Injectable()
@@ -56,9 +57,11 @@ export class ApplicationInitializeService {
     private readonly hashService = inject(HashService);
     private readonly toastService = inject(ToastService);
     private readonly store = inject(Store);
+    private readonly angulartics = inject(Angulartics2GoogleGlobalSiteTag);
 
     public async initialize() {
         try {
+            this.angulartics.startTracking();
             await this.loggingService.initialize();
             await this.loggingService.info("---------------------------------------");
             await this.loggingService.info("Starting IHM Application Initialization");
